fix session information
