+++ app/controllers/base_controller.rb
@@ -15,7 +15,11 @@ class BaseController < ApplicationController
   # skip_before_action :ensure_tenanted_user
   #
   include Authentication
-  before_action :check_session_length
+  #
+  # This next include makes sure
+  # that users do sign in at least every
+  # 7 days
+  include SessionGuard
   #
   #
   #
@@ -26,13 +30,4 @@ class BaseController < ApplicationController
   # def pagy_get_count(collection, vars)
   #   collection.respond_to?(:count_documents) ? collection.count_documents : super
   # end
-
-  def check_session_length
-    return unless user_signed_in?
-
-    if Time.now > Current.user.last_sign_in_at + 7.days
-      sign_out(current_user)
-      redirect_to new_user_session_path, warning: I18n.t("session_expired")
-    end
-  end
 end
